bridge rust code compiles (without tests)
