+++ bridge/src/lib.rs
@@ -105,8 +105,8 @@ pub use ordered_stream::OrderedStream;
 pub mod helpers;
 mod main_contract;
 pub use main_contract::MainContract;
-mod main_to_side_sign;
-pub use main_to_side_sign::MainToSideSign;
+mod accept_message_from_main;
+pub use accept_message_from_main::AcceptMessageFromMain;
 mod relay_stream;
 pub use relay_stream::RelayStream;
 mod send_tx_with_receipt;
